refactor(app): split near-cap SPEC AC barrels under size gate
Extract shared combat/move/dialogue frame helpers into app/test/support
and split combat_ui / move_dialogs / dialogue_overlays SPEC barrels into
part suites so wave-4 mega-suite densify continues without behavior change
(Refs #4013).
