+++ SPEC/ui/components/move-units-dialog-base.md
@@ -110,7 +110,7 @@ The trailing `Wrap` lets Cancel/Confirm flow onto a second run at narrow viewpor
 ## Tests
 
 - `app/test/move_units_dialog_base_test.dart` — widget-level contract tests pinning the idle/selected outline widths, the radio-dot fill, the trailing slot, and the tap callback.
-- `app/test/move_dialogs_specs_test.dart` — pins both consumer dialogs over this chrome (section labels, confirm/cancel, declare-war flow, locate).
+- `app/test/move_dialogs_specs_part1_test.dart` / `move_dialogs_specs_part2_test.dart` — pin army/fleet consumer dialogs over this chrome (section labels, confirm/cancel, declare-war flow, locate).
 - `app/test/move_dialogs_320dp_min_viewport_test.dart` — pins both dialogs at `kMinViewportWidth = 320` dp without overflow.
 
 ---
